Ironed out some handlebars / css bugs and corrected the POST register route
