+++ ProjectSourceCode/handlebars/views/pages/home.hbs
@@ -1,13 +1,5 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>{{title}}</title>
-  <link rel="stylesheet" href="/style.css">
-</head>
-<body>
-
+{{> head }}
+{{> siteheader }}
   <section class="hero">
     <h1>CU Marketplace</h1>
     <p class="hero-tagline">
@@ -43,10 +35,4 @@
       <li><strong>Connect</strong> with other students to make a deal.</li>
     </ol>
   </section>
-
-  <footer class="site-footer">
-    <p>&copy; 2024 CU Marketplace. All rights reserved.</p>
-  </footer>
-
-</body>
-</html>
+  {{>footer}}
