+++ ProjectSourceCode/handlebars/views/pages/login.hbs
@@ -1,36 +1,38 @@
 {{> head }}
 {{> siteheader }}
-<h2>Welcome Back! Please Login</h2>
+{{> message}}
 
-<form action="/api/users/login" method="POST" class="w-50">
-  <div class="mb-3">
-    <label for="username" class="form-label">Username</label>
-    <input 
-      type="text" 
-      class="form-control" 
-      id="username" 
-      name="username" 
-      placeholder="Enter username" 
-      required
-    >
-  </div>
+<section class="section">
+    <h2 class="text-center">Welcome Back! Please Login</h2>
 
-  <div class="mb-3">
-    <label for="password" class="form-label">Password</label>
-    <input 
-      type="password" 
-      class="form-control" 
-      id="password" 
-      name="password" 
-      placeholder="Enter password" 
-      required
-    >
-  </div>
+  <form action="/api/users/login" method="POST" class="auth-form">
+    <div class="form-group">
+      <label for="username">Username</label>
+      <input 
+        type="text" 
+        id="username" 
+        name="username" 
+        placeholder="Enter username" 
+        required
+      >
+    </div>
 
-  <button type="submit" class="btn btn-primary">Login</button>
-
-  <p class="mt-3">
-    <a href="/register">Register here!</a>
-  </p>
-</form>
+    <div class="form-group">
+      <label for="password">Password</label>
+      <input 
+        type="password" 
+        id="password" 
+        name="password" 
+        placeholder="Enter password" 
+        required
+      >
+    </div>
+    <div class="form-actions">
+      <button type="submit" class="btn">Login</button>
+    </div>
+    <p class="form-footer">
+      <a href="/register">Register here!</a>
+    </p>
+  </form>
+</section>
 {{>footer}}
