+++ ProjectSourceCode/handlebars/views/pages/post_card.hbs
@@ -1,53 +1,31 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>{{title}}</title>
-  <link rel="stylesheet" href="/style.css">
-</head>
-<body>
+{{> head }}
+{{> siteheader }}
+{{> message}}
 
-  <nav class="site-nav">
-    <div class="site-nav-inner">
-      <a href="/" class="site-brand">CU Marketplace</a>
-      <div class="site-nav-links">
-        <a href="/browse">Browse</a>
-        <a href="/post">Post</a>
-        <a href="/login">Login</a>
-        <a href="/register">Register</a>
-      </div>
-    </div>
-  </nav>
+<section class="section">
+  <h2 class="section-title">Available Items</h2>
 
-  {{> message}}
-
-  <section class="section">
-    <h2 class="section-title">Available Items</h2>
-
-    <div class="post-grid">
-      {{#each results}}
-        <div class="post-card">
-          <img
-            src="{{#if images.[0].url}}{{images.[0].url}}{{else}}https://via.placeholder.com/400x250?text=No+Image+Available{{/if}}"
-            alt="{{name}}"
-            class="post-card-image"
-          >
-          <div class="post-card-content">
-            <h3 class="post-card-title">{{name}}</h3>
-            <p class="post-card-description">
-              {{#if product.info}}
-                <strong>Desc:</strong> {{product.info}}<br>
-              {{/if}}
-              {{#if user.contact}}
-                <strong>Contact:</strong> {{user.contact}}
-              {{/if}}
-            </p>
-          </div>
+  <div class="post-grid">
+    {{#each results}}
+      <div class="post-card">
+        <img
+          src="{{#if images.[0].url}}{{images.[0].url}}{{else}}https://via.placeholder.com/400x250?text=No+Image+Available{{/if}}"
+          alt="{{name}}"
+          class="post-card-image"
+        >
+        <div class="post-card-content">
+          <h3 class="post-card-title">{{name}}</h3>
+          <p class="post-card-description">
+            {{#if product.info}}
+              <strong>Desc:</strong> {{product.info}}<br>
+            {{/if}}
+            {{#if user.contact}}
+              <strong>Contact:</strong> {{user.contact}}
+            {{/if}}
+          </p>
         </div>
-      {{/each}}
-    </div>
-  </section>
-
-</body>
-</html>
+      </div>
+    {{/each}}
+  </div>
+</section>
+{{>footer}}
