+++ ProjectSourceCode/handlebars/views/pages/register.hbs
@@ -1,31 +1,11 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>{{title}}</title>
-  <link rel="stylesheet" href="/style.css">
-</head>
-<body>
-
-  <nav class="site-nav">
-    <div class="site-nav-inner">
-      <a href="/" class="site-brand">CU Marketplace</a>
-      <div class="site-nav-links">
-        <a href="/browse">Browse</a>
-        <a href="/post">Post</a>
-        <a href="/login">Login</a>
-        <a href="/register">Register</a>
-      </div>
-    </div>
-  </nav>
-
-  {{> message}}
+{{> head }}
+{{> siteheader }}
+{{> message}}
 
   <section class="section">
     <h2 class="text-center">Register (CU Student Email Required)</h2>
 
-    <form action="/register" method="POST" class="auth-form">
+    <form action="/api/users/register" method="POST" class="auth-form">
       <div class="form-group">
         <label for="username">Username</label>
         <input 
@@ -62,7 +42,6 @@
 
       <div class="form-actions">
         <button type="submit" class="btn">Register</button>
-        <a href="/login" class="btn btn-secondary">Back to Login</a>
       </div>
 
       <p class="form-footer">
@@ -70,6 +49,4 @@
       </p>
     </form>
   </section>
-
-</body>
-</html>
+  {{>footer}}
