+++ ProjectSourceCode/handlebars/views/partials/head.hbs
@@ -9,6 +9,6 @@
   <meta name="description" content="" />
   {{> title }}
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-
+  <link rel="stylesheet" href="/style.css">
 </head>
 <body class="h-100 d-flex flex-column">
