--- ProjectSourceCode/index.js
@@ -1,213 +1,0 @@
-// index.js
-import express from "express";
-import { engine } from "express-handlebars";
-
-const app = express();
-const PORT = 4444;
-
-// Log incoming requests so we see what's happening
-app.use((req, res, next) => {
-  console.log("➡️", req.method, req.url);
-  next();
-});
-
-// Handlebars setup
-app.engine("hbs", engine({ extname: ".hbs", partialsDir: "./ProjectSourceCode/handlebars/views/partials" })); //tell handlebars where to find partials
-app.set("view engine", "hbs");
-app.set("views", "./ProjectSourceCode/handlebars/views/pages");
-
-// Parse form submissions (for later)
-app.use(express.urlencoded({ extended: true }));
-
-// ---------- ROUTES ----------
-
-// HOME – this is what / should show
-app.get("/", (req, res) => {
-  console.log("Rendering home.hbs");
-  res.render("home", {
-    layout: false,
-    title: "CU Marketplace",
-    loggedIn: false,
-    username: "Mike",
-  });
-});
-
-
-// LOGIN
-app.get("/login", (req, res) => {
-  console.log("Rendering login.hbs");
-  res.render("login", {
-    layout: false,
-    title: "Login - CU Marketplace",
-  });
-});
-
-// REGISTER
-app.get("/register", (req, res) => {
-  console.log("Rendering register.hbs");
-  res.render("register", {
-    layout: false,
-    title: "Register - CU Marketplace",
-  });
-});
-
-// POST ITEM
-app.get("/post", (req, res) => {
-  console.log("Rendering post_card.hbs");
-  
-  //temporary mock data for testing
-  const sampleResults = [{
-    name: "CSCI 2270 Textbook",
-      product: { info: "Gently used, 8/10 condition" },
-      user: { contact: "mike@colorado.edu" },
-      images: [{ url: "https://via.placeholder.com/400x250?text=Textbook" }]
-    },
-    {
-      name: "Mini Fridge",
-      product: { info: "Works perfectly, pickup only" },
-      user: { contact: "student1@colorado.edu" },
-      images: [{ url: "https://via.placeholder.com/400x250?text=Mini+Fridge" }]
-    },
-    {
-      name: "TI-84 Calculator",
-      product: { info: "Like new, includes batteries" },
-      user: { contact: "jane@colorado.edu" },
-      images: [{ url: "https://via.placeholder.com/400x250?text=Calculator" }]
-  }];
-
-  res.render("post_card", {
-    layout: false,
-    title: "Post an Item - CU Marketplace",
-    results: sampleResults,
-    message: "Sample items loaded successfully.",
-  });
-});
-
-// ---------- STATIC FILES (CSS, images, etc.) ----------
-
-// This MUST be after the routes so it doesn't override "/"
-app.use(express.static("ProjectSourceCode"));
-
-// ------------------------------
-
-app.listen(PORT, () => {
-  console.log(`✅ Server running at http://localhost:${PORT}`);
-});
-
-
-
-
-
-////////////////////
-// USER ENDPOINTS //
-////////////////////
-
-// create a new user
-app.post("/api/users/register", (req, res) => {
-
-});
-
-// authenticate user and return token/session
-app.post("/api/users/login", (req, res) => {
-
-});
-
-// log user out
-app.post("/api/users/logout", (req, res) => {
-
-});
-
-// return all users
-app.get("/api/users", (req, res) => {
-
-});
-
-// get user by id
-app.get("/api/users/:userId", (req, res) => {
-
-});
-
-
-
-////////////////////
-// POST ENDPOINTS //
-////////////////////
-
-// get all posts
-app.get("/api/posts", (req, res) => {
-
-});
-
-// get post by id
-app.get("/api/posts/:postId", (req, res) => {
-
-});
-
-// get posts by user id
-app.get("/api/posts/user/:userId", (req, res) => {
-
-});
-
-
-// create post
-app.post("/api/posts", (req, res) => {
-
-});
-
-// edit post by id
-app.put("/api/posts/:postId", (req, res) => {
-
-});
-
-
-// delete post by id
-app.delete("/api/posts/:postId", (req, res) => {
-
-});
-
-
-// get posts by search keyword
-app.get("/api/posts/search", (req, res) => {
-  const keyword = req.query.q;
-
-  if (!keyword) {
-    return res.status(400).json({ error: "Missing search keyword 'q' parameter" });
-  }
-
-  res.json({ message: `Searching posts with keyword: ${keyword}` });
-});
-
-// get posts by category name
-app.get("/api/posts/category/:name", (req, res) => {
-
-});
-
-// update post status (available, sold)
-app.patch("/api/posts/:postId/status", (req, res) => {
-
-});
-
-
-////////////////////
-// CATEGORY ENDPOINTS
-////////////////////
-
-// get all categories
-app.get("/api/categories", (req, res) => {
-
-});
-
-// get category by id
-app.get("/api/categories/:categoryId", (req, res) => {
-
-});
-
-// create new category
-app.post("/api/categories", (req, res) => {
-
-});
-
-// update category by id
-app.delete("/api/categories/:categoryId", (req, res) => {
-
-});
