+++ package.json
@@ -33,7 +33,7 @@
     "handlebars": "^4.7.8",
     "nodemon": "^3.1.10",
     "pg-promise": "^12.2.0"
-    
+  },
   "devDependencies": {
     "chai": "^6.2.1",
     "chai-http": "^4.4.0",
